refactor llm preprocess and process for model_op before introduce llm rater
