+++ uniflow/constants.py
@@ -6,3 +6,4 @@ OUTPUT_NAME = "output"
 # Flow Types
 EXTRACT = "extract"
 TRANSFORM = "transform"
+CLASSIFICATION = "classification"
