+++ uniflow/flow/transform/transform_huggingface_flow.py
@@ -4,7 +4,8 @@ from typing import Any, Dict, Sequence
 from uniflow.constants import TRANSFORM
 from uniflow.flow.flow import Flow
 from uniflow.node import Node
-from uniflow.op.model.model_op import LLMDataProcessor, ModelOp
+from uniflow.op.model.llm_processor import LLMDataProcessor
+from uniflow.op.model.model_op import ModelOp
 from uniflow.op.prompt_schema import GuidedPrompt
