+++ uniflow/op/model/constants.py
@@ -1,0 +1,7 @@
+"""Constants for the op model."""
+
+RESPONSE = "response"
+ERROR = "error"
+ERROR_LIST = "error_list"
+ERROR_CONTEXT = "error_context"
+MAX_ATTEMPTS = 3
